requirements.txt
Django==4.0.6
mysql-connector==2.2.9
numpy==1.23.1
pandas==1.4.3
python-dotenv==0.19.1

